+++ src/Data/Containers/NonEmpty.hs
@@ -20,7 +20,7 @@
 --
 -- Used to convert between and in between possibly-empty and non-empty
 -- types.  Instances are provided for all modules in this package, as well
--- as for 'NonEmpty' in /base/.
+-- as for 'NonEmpty' in /base/ and 'NonEmptyVector'.
 module Data.Containers.NonEmpty (
     HasNonEmpty(..)
   , pattern IsNonEmpty, pattern IsEmpty